+++ bench/src/single_frida.rs
@@ -14,14 +14,14 @@ use crate::common::{field_names, Blake3F128, Blake3F64, F128Element, F64Element,
 #[derive(Debug)]
 pub struct SingleFridaBenchmarkResult {
     field_type: String,
-    batch_size: usize,
-    blowup_factor: usize,
-    folding_factor: usize,
-    max_remainder_degree: usize,
-    data_size_kb: usize,
+    batch_size: u32,
+    blowup_factor: u32,
+    folding_factor: u32,
+    max_remainder_degree: u32,
+    data_size_kb: u64,
     domain_size: usize,
     single_proof_time_ms: f64,
-    single_proof_size_bytes: usize,
+    single_proof_size_bytes: u64,
     total_proof_size_estimate_mb: f64,
 }
 
@@ -49,7 +49,7 @@ impl SingleFridaBenchmarkResult {
 
 fn benchmark_non_batched<E, H>(
     options: FriOptions,
-    data_size: usize,
+    data_size: u64,
     field_name: &str,
 ) -> SingleFridaBenchmarkResult
 where
@@ -59,14 +59,14 @@ where
     let mut total_proof_time = Duration::ZERO;
     let mut total_proof_size = 0;
 
-    let encoded_element_count = encoded_data_element_count::<E>(data_size);
+    let encoded_element_count = encoded_data_element_count::<E>(data_size as usize);
     let domain_size = usize::max(
         encoded_element_count.next_power_of_two() * options.blowup_factor(),
         constants::MIN_DOMAIN_SIZE,
     );
 
     for _ in 0..RUNS {
-        let data = rand_vector::<u8>(data_size);
+        let data = rand_vector::<u8>(data_size as usize);
         let prover_builder = FridaProverBuilder::<E, H>::new(options.clone());
 
         let (_, prover, base_positions) = prover_builder
@@ -78,20 +78,20 @@ where
         let start = Instant::now();
         let proof = prover.open(&drawn_position);
         total_proof_time += start.elapsed();
-        total_proof_size += proof.size();
+        total_proof_size += proof.size() as u64;
     }
 
     let avg_proof_time_ms = total_proof_time.as_secs_f64() * 1000.0 / RUNS as f64;
-    let avg_proof_size_bytes = total_proof_size / RUNS;
+    let avg_proof_size_bytes = total_proof_size / RUNS as u64;
     let total_proof_size_estimate_mb =
-        (domain_size * avg_proof_size_bytes) as f64 / (1024.0 * 1024.0);
+        (domain_size as u64 * avg_proof_size_bytes) as f64 / (1024.0 * 1024.0);
 
     SingleFridaBenchmarkResult {
         field_type: field_name.to_string(),
         batch_size: 1,
-        blowup_factor: options.blowup_factor(),
-        folding_factor: options.folding_factor(),
-        max_remainder_degree: options.remainder_max_degree(),
+        blowup_factor: options.blowup_factor() as u32,
+        folding_factor: options.folding_factor() as u32,
+        max_remainder_degree: options.remainder_max_degree() as u32,
         data_size_kb: data_size / 1024,
         domain_size,
         single_proof_time_ms: avg_proof_time_ms,
@@ -102,8 +102,8 @@ where
 
 fn benchmark_batched<E, H>(
     options: FriOptions,
-    data_size: usize,
-    batch_size: usize,
+    data_size: u64,
+    batch_size: u32,
     field_name: &str,
 ) -> SingleFridaBenchmarkResult
 where
@@ -113,7 +113,7 @@ where
     let mut total_proof_time = Duration::ZERO;
     let mut total_proof_size = 0;
 
-    let max_data_len = encoded_data_element_count::<E>(data_size);
+    let max_data_len = encoded_data_element_count::<E>(data_size as usize);
     let domain_size = usize::max(
         (max_data_len * options.blowup_factor()).next_power_of_two(),
         constants::MIN_DOMAIN_SIZE,
@@ -122,7 +122,7 @@ where
     for _ in 0..RUNS {
         let mut data_list = vec![];
         for _ in 0..batch_size {
-            data_list.push(rand_vector::<u8>(data_size));
+            data_list.push(rand_vector::<u8>(data_size as usize));
         }
 
         let prover_builder = FridaProverBuilder::<E, H>::new(options.clone());
@@ -136,20 +136,20 @@ where
         let start = Instant::now();
         let proof = prover.open(&drawn_position);
         total_proof_time += start.elapsed();
-        total_proof_size += proof.size();
+        total_proof_size += proof.size() as u64;
     }
 
     let avg_proof_time_ms = total_proof_time.as_secs_f64() * 1000.0 / RUNS as f64;
-    let avg_proof_size_bytes = total_proof_size / RUNS;
+    let avg_proof_size_bytes = total_proof_size / RUNS as u64;
     let total_proof_size_estimate_mb =
-        (domain_size * avg_proof_size_bytes) as f64 / (1024.0 * 1024.0);
+        (domain_size as u64 * avg_proof_size_bytes) as f64 / (1024.0 * 1024.0);
 
     SingleFridaBenchmarkResult {
         field_type: field_name.to_string(),
         batch_size,
-        blowup_factor: options.blowup_factor(),
-        folding_factor: options.folding_factor(),
-        max_remainder_degree: options.remainder_max_degree(),
+        blowup_factor: options.blowup_factor() as u32,
+        folding_factor: options.folding_factor() as u32,
+        max_remainder_degree: options.remainder_max_degree() as u32,
         data_size_kb: data_size / 1024,
         domain_size,
         single_proof_time_ms: avg_proof_time_ms,
